requirements.txt
matplotlib==3.7.1
numpy==1.24.2
opencv-python-headless
pandas==1.5.3
Pillow==9.4.0
pytesseract==0.3.10
PyYAML==6.0
scikit_learn==1.2.2
streamlit==1.20.0
tensorflow==2.11.0
gunicorn==20.1.0

